+++ src/model/config.rs
@@ -314,6 +314,8 @@ pub struct ConfigTargetOptions {
     pub m3u_mask_redirect_url: bool,
     #[serde(default)]
     pub share_live_streams: bool,
+    #[serde(default)]
+    pub remove_duplicates: bool,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]